+++ packages/ai/core/prompt/convert-to-core-messages.ts
@@ -1,4 +1,5 @@
 import {
+  FileUIPart,
   Message,
   ReasoningUIPart,
   TextUIPart,
@@ -40,15 +41,31 @@ export function convertToCoreMessages<TOOLS extends ToolSet = never>(
       }
 
       case 'user': {
-        coreMessages.push({
-          role: 'user',
-          content: experimental_attachments
-            ? [
-                { type: 'text', text: content },
-                ...attachmentsToParts(experimental_attachments),
-              ]
-            : content,
-        });
+        if (message.parts == null) {
+          coreMessages.push({
+            role: 'user',
+            content: experimental_attachments
+              ? [
+                  { type: 'text', text: content },
+                  ...attachmentsToParts(experimental_attachments),
+                ]
+              : content,
+          });
+        } else {
+          const textParts = message.parts
+            .filter(part => part.type === 'text')
+            .map(part => ({
+              type: 'text' as const,
+              text: part.text,
+            }));
+
+          coreMessages.push({
+            role: 'user',
+            content: experimental_attachments
+              ? [...textParts, ...attachmentsToParts(experimental_attachments)]
+              : textParts,
+          });
+        }
         break;
       }
 
@@ -57,7 +74,7 @@ export function convertToCoreMessages<TOOLS extends ToolSet = never>(
           let currentStep = 0;
           let blockHasToolInvocations = false;
           let block: Array<
-            TextUIPart | ToolInvocationUIPart | ReasoningUIPart
+            TextUIPart | ToolInvocationUIPart | ReasoningUIPart | FileUIPart
           > = [];
 
           function processBlock() {
@@ -65,12 +82,11 @@ export function convertToCoreMessages<TOOLS extends ToolSet = never>(
 
             for (const part of block) {
               switch (part.type) {
-                case 'text':
-                  content.push({
-                    type: 'text' as const,
-                    text: part.text,
-                  });
+                case 'file':
+                case 'text': {
+                  content.push(part);
                   break;
+                }
                 case 'reasoning': {
                   for (const detail of part.details) {
                     switch (detail.type) {
@@ -115,7 +131,11 @@ export function convertToCoreMessages<TOOLS extends ToolSet = never>(
             const stepInvocations = block
               .filter(
                 (
-                  part: TextUIPart | ToolInvocationUIPart | ReasoningUIPart,
+                  part:
+                    | TextUIPart
+                    | ToolInvocationUIPart
+                    | ReasoningUIPart
+                    | FileUIPart,
                 ): part is ToolInvocationUIPart =>
                   part.type === 'tool-invocation',
               )
@@ -167,13 +187,15 @@ export function convertToCoreMessages<TOOLS extends ToolSet = never>(
 
           for (const part of message.parts) {
             switch (part.type) {
-              case 'reasoning':
-                block.push(part);
-                break;
               case 'text': {
                 if (blockHasToolInvocations) {
                   processBlock(); // text must come before tool invocations
                 }
+                block.push(part);
+                break;
+              }
+              case 'file':
+              case 'reasoning': {
                 block.push(part);
                 break;
               }