+++ packages/ui-utils/src/types.ts
@@ -108,7 +108,12 @@ that the assistant made as part of this message.
    */
   // note: optional on the Message type (which serves as input)
   parts?: Array<
-    TextUIPart | ReasoningUIPart | ToolInvocationUIPart | SourceUIPart
+    | TextUIPart
+    | ReasoningUIPart
+    | ToolInvocationUIPart
+    | SourceUIPart
+    | FileUIPart
+    | StepStartUIPart
   >;
 }
 
@@ -120,7 +125,12 @@ export type UIMessage = Message & {
    * User messages can have text parts.
    */
   parts: Array<
-    TextUIPart | ReasoningUIPart | ToolInvocationUIPart | SourceUIPart
+    | TextUIPart
+    | ReasoningUIPart
+    | ToolInvocationUIPart
+    | SourceUIPart
+    | FileUIPart
+    | StepStartUIPart
   >;
 };
 
@@ -176,6 +186,22 @@ export type SourceUIPart = {
    * The source.
    */
   source: LanguageModelV1Source;
+};
+
+/**
+ * A file part of a message.
+ */
+export type FileUIPart = {
+  type: 'file';
+  mimeType: string;
+  data: string; // base64 encoded data
+};
+
+/**
+ * A step boundary part of a message.
+ */
+export type StepStartUIPart = {
+  type: 'step-start';
 };
 
 export type CreateMessage = Omit<Message, 'id'> & {